requirements.txt
numpy
pillow
requests
pyyaml
tqdm
streamlit
chromadb
sentence-transformers
scenedetect[opencv]
openai-whisper
easyocr
